isoprocessor2: Process Stable Isotope Data

Data processing and reduction pipelines for the stable isotope data read by the isoreader2 package, including peak detection, calibration, standardization, and visualization of isotope ratio mass spectrometry (IRMS) data.

Details

Resources:

Website for the isoprocessor2 package: https://isoprocessor2.isoverse.org

Package options: ip_options

See also

Author

Maintainer: Sebastian Kopf sebastian.kopf@colorado.edu (ORCID)

Authors:

Sebastian Kopf sebastian.kopf@colorado.edu (ORCID)